+++ src/api/instance/copy.js
@@ -4,6 +4,7 @@ const Promise = require('bluebird');
 const AWSProvider = require('../aws_provider');
 const AWSUtil = require('../aws_util');
 const logger = require('../../logger');
+const Override = require('../override');
 
 module.exports = function (regions, instance_name, rename, vpc, ami, i_type, key_name, sg, iam, ud_files, rud_file, raw_ud_string, disks, az, tags, eni_name, env_vars, ebs_opt) {
 	if(!ud_files) ud_files = [];
@@ -38,7 +39,7 @@ function copy(region, instance_name, rename, override_opts) {
 			return copy_instance_attrs(region, instance.InstanceId);
 		})
 		.then(function (instance_attrs) {
-			return merge_config(region, instance_attrs, override_opts).then(merged => {
+			return Override.merge_config(region, instance_attrs, override_opts).then(merged => {
 				return rename_instance(merged, rename);
 			});
 		})
@@ -50,50 +51,6 @@ function copy(region, instance_name, rename, override_opts) {
 			logger.info('Wait for instance intialization');
 			return wait_until_runnning(region, data.Instances[0].InstanceId);
 		});
-}
-
-function _build_override_params(region, oo) {
-	return Promise.all([
-			AWSUtil.get_ami_id(region, oo.ami),
-			AWSUtil.get_userdata_string(oo.ud_files, oo.env_vars, oo.raw_ud_string),
-			AWSUtil.get_network_interface(region, oo.vpc, oo.az, oo.eni_name, oo.sg),
-			AWSUtil.get_bdms(oo.disks)
-		])
-		.spread(function(ami_id, userdata_string, network_interface, bdms){
-			let config = {
-				BlockDeviceMappings: bdms,
-				ImageId: ami_id,
-				InstanceType: oo.i_type,
-				KeyName: oo.key_name,
-				NetworkInterfaces: network_interface,
-			};
-			if (typeof(oo.ebs_opt) === 'boolean') {
-				config.EbsOptimized = !!oo.ebs_opt;
-			}
-			if (oo.iam) {
-				config.IamInstanceProfile = {
-					Name: oo.iam
-				};
-			}
-			if ((oo.ud_files && oo.ud_files.length) || oo.raw_ud_string) {
-				config.UserData = new Buffer(userdata_string).toString('base64');
-			}
-			return config;
-		});
-}
-
-function merge_config(region, orig_config, override_opts) {
-	return _build_override_params(region, override_opts).then(function(override_config) {
-		for (let key of Object.keys(override_config)) {
-			let val = override_config[key];
-			// Don't assign null, empty or undefined configurations
-			if (val == null || (Array.isArray(val) && !val.length)) {
-				delete override_config[key];
-			}
-		}
-		let new_config = Object.assign(orig_config, override_config);
-		return new_config;
-	});
 }
 
 function wait_until_runnning(region, instanceId) {
